Added Insert/Delete methods to DataEntry app
Added methods for Insert/Delete and changed the UI to accommodate three radio buttons (Insert/Modify/Delete) with the latter 2 requiring manual mRID entry.  Also tweaked the appearance of the GUI to be more symmetrical and changed it so that it would launch in the center of the screen instead of the corner.